% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceBC.R
\name{PriceBC}
\alias{PriceBC}
\title{Calculate bills after taxes and tips}
\usage{
PriceBC(type, PriceTag, MealBill, AlcoholBill, Tip)
}
\arguments{
\item{type}{Character vector defining the type of service, food or retail. Taxes and tips apply differently to each service. Requires "Restaurant" or "Sale" as argument input.}

\item{PriceTag}{Numeric vector indicating the cost of products to buy in the retail service, before taxes.}

\item{MealBill}{Numeric vector indicating the cost of food and non-alcoholic beverages in the food service, before taxes.}

\item{AlcoholBill}{Numeric vector indicating the cost of alcoholic beverages, before taxes.}

\item{Tip}{Numeric vector indicating the percentage of the total bill to give as tip.}
}
\value{
A vector displaying the final bill after taxes and after tip.
}
\description{
Calculates the additional costs of taxes and tips when buying in the food and retail services in British Columbia, Canada.
}
\examples{
PriceBC("Sale", PriceTag = 50)
PriceBC("Restaurant", MealBill = 35, AlcoholBill = 67, Tip = 18)

}
